rename `spatial_repr` module to `geometry`
