+++ honeycomb-core/src/cmap/dim2/advanced_ops.rs
@@ -5,8 +5,8 @@
 
 // ------ IMPORTS
 
+use crate::geometry::CoordsFloat;
 use crate::prelude::{CMap2, DartIdentifier, EdgeIdentifier, Vertex2, NULL_DART_ID};
-use crate::spatial_repr::CoordsFloat;
 
 // ------ CONTENT
