+++ honeycomb-core/src/cmap/dim2/embed.rs
@@ -9,7 +9,7 @@
 use crate::prelude::{AttributeBind, AttributeUpdate, CMap2, CMapError, Vertex2, VertexIdentifier};
 use crate::{
     attributes::{AttributeStorage, UnknownAttributeStorage},
-    spatial_repr::CoordsFloat,
+    geometry::CoordsFloat,
 };
 
 // ------ CONTENT
