+++ honeycomb-core/src/prelude.rs
@@ -6,7 +6,7 @@ pub use crate::cmap::{
     Orbit2, OrbitPolicy, VertexIdentifier, VolumeIdentifier, NULL_DART_ID, NULL_EDGE_ID,
     NULL_FACE_ID, NULL_VERTEX_ID, NULL_VOLUME_ID,
 };
-pub use crate::spatial_repr::{CoordsError, Vector2, Vertex2};
+pub use crate::geometry::{CoordsError, Vector2, Vertex2};
 
 // ------ FEATURE-GATED RE-EXPORTS
